V6 = V5 + fixtures

+++ src/Entity/Cours.php
@@ -23,7 +23,7 @@ class Cours
     private ?string $description = null;
 
     #[ORM\Column]
-    private ?int $prix = null;
+    private ?float $prix = null;
 
     #[ORM\Column]
     private ?bool $esCollectif = null;
@@ -78,12 +78,12 @@ class Cours
         return $this;
     }
 
-    public function getPrix(): ?int
+    public function getPrix(): ?float
     {
         return $this->prix;
     }
 
-    public function setPrix(int $prix): static
+    public function setPrix(float $prix): static
     {
         $this->prix = $prix;
